qa: move `laminas-cache` integration tests to own namespace
