+++ test/integration/Laminas/RedisClusterStorageCreationTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace LaminasTest\Cache\Storage\Adapter\Redis;
+namespace LaminasTest\Cache\Storage\Adapter\Laminas;
 
 use Laminas\Cache\Storage\Adapter\RedisCluster;
 use Laminas\Cache\Storage\Adapter\RedisClusterOptions;
