+++ test/integration/Laminas/RedisStorageCreationTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace LaminasTest\Cache\Storage\Adapter\Redis;
+namespace LaminasTest\Cache\Storage\Adapter\Laminas;
 
 use Laminas\Cache\Storage\Adapter\Redis;
 use Laminas\Cache\Storage\Adapter\RedisOptions;
